+++ pkgs/cargoHelperFunctions.sh
@@ -6,3 +6,9 @@ cargo() {
     command cargo "$@"
   )
 }
+
+# Injects `--profile $CARGO_PROFILE` into a particular cargo invocation
+# if the environment variable is set
+cargoWithProfile() {
+  cargo "${@:1:1}" ${CARGO_PROFILE:+--profile ${CARGO_PROFILE}} "${@:2}"
+}